requirements.txt
eventlet==0.24.1
flask==1.0.2
flask-socketio==3.1.1
python-socketio==2.1.2
